+++ db/db_block_cache_test.cc
@@ -250,7 +250,7 @@ TEST_F(DBBlockCacheTest, TestWithoutCompressedBlockCache) {
   cache->SetStrictCapacityLimit(true);
   iter = db_->NewIterator(read_options);
   iter->Seek(std::to_string(kNumBlocks - 1));
-  ASSERT_TRUE(iter->status().IsIncomplete());
+  ASSERT_TRUE(iter->status().IsMemoryLimit());
   CheckCacheCounters(options, 1, 0, 0, 1);
   delete iter;
   iter = nullptr;
@@ -333,8 +333,10 @@ TEST_F(DBBlockCacheTest, TestWithCompressedBlockCache) {
   ASSERT_EQ(usage, cache->GetPinnedUsage());
 
   // Load last key block.
-  ASSERT_EQ("Result incomplete: Insert failed due to LRU cache being full.",
-            Get(std::to_string(kNumBlocks - 1)));
+  ASSERT_EQ(
+      "Operation aborted: Memory limit reached: Insert failed due to LRU cache "
+      "being full.",
+      Get(std::to_string(kNumBlocks - 1)));
   // Failure will also record the miss counter.
   CheckCacheCounters(options, 1, 0, 0, 1);
   CheckCompressedCacheCounters(options, 1, 0, 1, 0);
@@ -398,7 +400,7 @@ class PersistentCacheFromCache : public PersistentCache {
   bool read_only_;
 };
 
-class ReadOnlyCacheWrapper : public CacheWrapper {
+class ReadOnlyCacheWrapper : public test::CacheWrapper {
   using CacheWrapper::CacheWrapper;
 
   using Cache::Insert;
@@ -901,7 +903,7 @@ namespace {
 // This allows us to manipulate BlockBasedTableReader into thinking
 // another thread inserted the data in between Lookup and Insert,
 // while mostly preserving the LRUCache interface/behavior.
-class LookupLiarCache : public CacheWrapper {
+class LookupLiarCache : public test::CacheWrapper {
   int nth_lookup_not_found_ = 0;
 
  public: