+++ db_stress_tool/db_stress_test_base.cc
@@ -8,6 +8,9 @@
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 //
 
+#include <exception>
+#include <memory>
+
 #include "util/compression.h"
 #ifdef GFLAGS
 #include "cache/fast_lru_cache.h"
@@ -15,6 +18,7 @@
 #include "db_stress_tool/db_stress_compaction_filter.h"
 #include "db_stress_tool/db_stress_driver.h"
 #include "db_stress_tool/db_stress_table_properties_collector.h"
+#include "port/stack_trace.h"
 #include "rocksdb/convenience.h"
 #include "rocksdb/filter_policy.h"
 #include "rocksdb/secondary_cache.h"
@@ -49,7 +53,7 @@ std::shared_ptr<const FilterPolicy> CreateFilterPolicy() {
 }  // namespace
 
 StressTest::StressTest()
-    : cache_(NewCache(FLAGS_cache_size, FLAGS_cache_numshardbits)),
+    : cache_(CreateCache()),
       compressed_cache_(NewLRUCache(FLAGS_compressed_cache_size,
                                     FLAGS_compressed_cache_numshardbits)),
       filter_policy_(CreateFilterPolicy()),
@@ -106,28 +110,30 @@ StressTest::~StressTest() {
   delete cmp_db_;
 }
 
-std::shared_ptr<Cache> StressTest::NewCache(size_t capacity,
-                                            int32_t num_shard_bits) {
-  ConfigOptions config_options;
-  if (capacity <= 0) {
+namespace {
+std::shared_ptr<Cache> CreateBaseCache() {
+  if (FLAGS_cache_size <= 0) {
     return nullptr;
   }
+  const size_t capacity = static_cast<size_t>(FLAGS_cache_size);
 
+  ConfigOptions config_options;
   if (FLAGS_cache_type == "clock_cache") {
-    auto cache = NewClockCache((size_t)capacity);
+    auto cache = NewClockCache(capacity);
     if (!cache) {
       fprintf(stderr, "Clock cache not supported.");
       exit(1);
     }
     return cache;
   } else if (FLAGS_cache_type == "fast_lru_cache") {
-    return NewFastLRUCache(static_cast<size_t>(capacity), FLAGS_block_size,
-                           num_shard_bits, false /*strict_capacity_limit*/,
+    return NewFastLRUCache(capacity, FLAGS_block_size, FLAGS_cache_numshardbits,
+                           false /*strict_capacity_limit*/,
                            kDefaultCacheMetadataChargePolicy);
   } else if (FLAGS_cache_type == "lru_cache") {
     LRUCacheOptions opts;
     opts.capacity = capacity;
-    opts.num_shard_bits = num_shard_bits;
+    opts.num_shard_bits = FLAGS_cache_numshardbits;
+    opts.strict_capacity_limit = FLAGS_cache_strict_capacity_limit;
 #ifndef ROCKSDB_LITE
     std::shared_ptr<SecondaryCache> secondary_cache;
     if (!FLAGS_secondary_cache_uri.empty()) {
@@ -151,6 +157,68 @@ std::shared_ptr<Cache> StressTest::NewCache(size_t capacity,
   } else {
     fprintf(stderr, "Cache type not supported.");
     exit(1);
+    return nullptr;
+  }
+}
+}  // namespace
+
+std::shared_ptr<Cache> StressTest::CreateCache() {
+  class StressTestCacheWrapper : public test::CacheWrapper {
+   public:
+    StressTestCacheWrapper(StressTest* stress_test,
+                           std::shared_ptr<Cache> target)
+        : CacheWrapper(std::move(target)),
+          cache_failed_insertion_balance_(
+              stress_test->cache_failed_insertion_balance_) {}
+
+    Status Insert(const Slice& key, void* value, size_t charge,
+                  DeleterFn deleter, Handle** handle = nullptr,
+                  Priority priority = Priority::LOW) override {
+      Status s = test::CacheWrapper::Insert(key, value, charge, deleter, handle,
+                                            priority);
+      if (s.IsMemoryLimit()) {
+        ++cache_failed_insertion_balance_;
+      }
+      return s;
+    }
+    Status Insert(const Slice& key, void* value, const CacheItemHelper* helper,
+                  size_t charge, Handle** handle = nullptr,
+                  Priority priority = Priority::LOW) override {
+      Status s = test::CacheWrapper::Insert(key, value, helper, charge, handle,
+                                            priority);
+      if (s.IsMemoryLimit()) {
+        ++cache_failed_insertion_balance_;
+      }
+      return s;
+    }
+
+   private:
+    std::atomic<int64_t>& cache_failed_insertion_balance_;
+  };
+
+  std::shared_ptr<Cache> base_cache = CreateBaseCache();
+  if (base_cache) {
+    return std::make_shared<StressTestCacheWrapper>(this, base_cache);
+  } else {
+    return nullptr;
+  }
+}
+
+void StressTest::VerifyOneOpMemoryLimit(const Status& s, ThreadState* thread) {
+  assert(s.IsMemoryLimit());
+  auto after = --cache_failed_insertion_balance_;
+  if (after < 0) {
+    if (thread) {
+      thread->shared->GetMutex()->Lock();
+    }
+    fprintf(stderr,
+            "Hit MemoryLimit error without matching Cache Insert failure: %s\n",
+            s.ToString().c_str());
+    port::PrintStack();
+    std::terminate();
+  } else {
+    // TODO stat in thread state?
+    (void)thread;
   }
 }
 
@@ -705,7 +773,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 
       if (thread->rand.OneInOpt(FLAGS_sync_wal_one_in)) {
         Status s = db_->SyncWAL();
-        if (!s.ok() && !s.IsNotSupported()) {
+        if (s.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(s, thread);
+        } else if (!s.ok() && !s.IsNotSupported()) {
           fprintf(stderr, "SyncWAL() failed: %s\n", s.ToString().c_str());
         }
       }
@@ -738,7 +808,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 
       if (thread->rand.OneInOpt(FLAGS_flush_one_in)) {
         Status status = TestFlush(rand_column_families);
-        if (!status.ok()) {
+        if (status.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(status, thread);
+        } else if (!status.ok()) {
           fprintf(stdout, "Unable to perform Flush(): %s\n",
                   status.ToString().c_str());
         }
@@ -749,7 +821,9 @@ void StressTest::OperateDb(ThreadState* thread) {
       if (thread->rand.OneInOpt(FLAGS_get_live_files_one_in) &&
           !FLAGS_write_fault_one_in) {
         Status status = VerifyGetLiveFiles();
-        if (!status.ok()) {
+        if (status.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(status, thread);
+        } else if (!status.ok()) {
           VerificationAbort(shared, "VerifyGetLiveFiles status not OK", status);
         }
       }
@@ -757,7 +831,9 @@ void StressTest::OperateDb(ThreadState* thread) {
       // Verify GetSortedWalFiles with a 1 in N chance.
       if (thread->rand.OneInOpt(FLAGS_get_sorted_wal_files_one_in)) {
         Status status = VerifyGetSortedWalFiles();
-        if (!status.ok()) {
+        if (status.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(status, thread);
+        } else if (!status.ok()) {
           VerificationAbort(shared, "VerifyGetSortedWalFiles status not OK",
                             status);
         }
@@ -766,7 +842,9 @@ void StressTest::OperateDb(ThreadState* thread) {
       // Verify GetCurrentWalFile with a 1 in N chance.
       if (thread->rand.OneInOpt(FLAGS_get_current_wal_file_one_in)) {
         Status status = VerifyGetCurrentWalFile();
-        if (!status.ok()) {
+        if (status.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(status, thread);
+        } else if (!status.ok()) {
           VerificationAbort(shared, "VerifyGetCurrentWalFile status not OK",
                             status);
         }
@@ -775,7 +853,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 
       if (thread->rand.OneInOpt(FLAGS_pause_background_one_in)) {
         Status status = TestPauseBackground(thread);
-        if (!status.ok()) {
+        if (status.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(status, thread);
+        } else if (!status.ok()) {
           VerificationAbort(
               shared, "Pause/ContinueBackgroundWork status not OK", status);
         }
@@ -784,7 +864,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 #ifndef ROCKSDB_LITE
       if (thread->rand.OneInOpt(FLAGS_verify_checksum_one_in)) {
         Status status = db_->VerifyChecksum();
-        if (!status.ok()) {
+        if (status.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(status, thread);
+        } else if (!status.ok()) {
           VerificationAbort(shared, "VerifyChecksum status not OK", status);
         }
       }
@@ -814,7 +896,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 
         if (total_size <= FLAGS_backup_max_size) {
           Status s = TestBackupRestore(thread, rand_column_families, rand_keys);
-          if (!s.ok()) {
+          if (s.IsMemoryLimit()) {
+            VerifyOneOpMemoryLimit(s, thread);
+          } else if (!s.ok()) {
             VerificationAbort(shared, "Backup/restore gave inconsistent state",
                               s);
           }
@@ -823,7 +907,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 
       if (thread->rand.OneInOpt(FLAGS_checkpoint_one_in)) {
         Status s = TestCheckpoint(thread, rand_column_families, rand_keys);
-        if (!s.ok()) {
+        if (s.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(s, thread);
+        } else if (!s.ok()) {
           VerificationAbort(shared, "Checkpoint gave inconsistent state", s);
         }
       }
@@ -832,7 +918,9 @@ void StressTest::OperateDb(ThreadState* thread) {
       if (thread->rand.OneInOpt(FLAGS_approximate_size_one_in)) {
         Status s =
             TestApproximateSize(thread, i, rand_column_families, rand_keys);
-        if (!s.ok()) {
+        if (s.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(s, thread);
+        } else if (!s.ok()) {
           VerificationAbort(shared, "ApproximateSize Failed", s);
         }
       }
@@ -843,7 +931,9 @@ void StressTest::OperateDb(ThreadState* thread) {
 
       /*always*/ {
         Status s = MaybeReleaseSnapshots(thread, i);
-        if (!s.ok()) {
+        if (s.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(s, thread);
+        } else if (!s.ok()) {
           VerificationAbort(shared, "Snapshot gave inconsistent state", s);
         }
       }
@@ -1163,6 +1253,9 @@ Status StressTest::TestIterate(ThreadState* thread,
 
     if (s.ok()) {
       thread->stats.AddIterations(1);
+    } else if (s.IsMemoryLimit()) {
+      VerifyOneOpMemoryLimit(s, thread);
+      break;
     } else {
       fprintf(stderr, "TestIterate error: %s\n", s.ToString().c_str());
       thread->stats.AddErrors(1);
@@ -2027,7 +2120,10 @@ Status StressTest::MaybeReleaseSnapshots(ThreadState* thread, uint64_t i) {
     db_->ReleaseSnapshot(snap_state.snapshot);
     delete snap_state.key_vec;
     thread->snapshot_queue.pop();
-    if (!s.ok()) {
+    if (s.IsMemoryLimit()) {
+      VerifyOneOpMemoryLimit(s, thread);
+      // Pretend everything OK
+    } else if (!s.ok()) {
       return s;
     }
   }
@@ -2076,21 +2172,30 @@ void StressTest::TestCompactRange(ThreadState* thread, int64_t rand_key,
     // Do some validation by declaring a snapshot and compare the data before
     // and after the compaction
     pre_snapshot = db_->GetSnapshot();
-    pre_hash =
+    auto pre_opt_hash =
         GetRangeHash(thread, pre_snapshot, column_family, start_key, end_key);
+    if (pre_opt_hash.has_value()) {
+      pre_hash = *pre_opt_hash;
+    } else {
+      // Tolerable failure, e.g. due to memory limit
+      db_->ReleaseSnapshot(pre_snapshot);
+      return;
+    }
   }
 
   Status status = db_->CompactRange(cro, column_family, &start_key, &end_key);
 
-  if (!status.ok()) {
+  if (status.IsMemoryLimit()) {
+    VerifyOneOpMemoryLimit(status, thread);
+  } else if (!status.ok()) {
     fprintf(stdout, "Unable to perform CompactRange(): %s\n",
             status.ToString().c_str());
   }
 
   if (pre_snapshot != nullptr) {
-    uint32_t post_hash =
+    auto post_opt_hash =
         GetRangeHash(thread, pre_snapshot, column_family, start_key, end_key);
-    if (pre_hash != post_hash) {
+    if (post_opt_hash.has_value() && pre_hash != *post_opt_hash) {
       fprintf(stderr,
               "Data hash different before and after compact range "
               "start_key %s end_key %s\n",
@@ -2103,10 +2208,10 @@ void StressTest::TestCompactRange(ThreadState* thread, int64_t rand_key,
   }
 }
 
-uint32_t StressTest::GetRangeHash(ThreadState* thread, const Snapshot* snapshot,
-                                  ColumnFamilyHandle* column_family,
-                                  const Slice& start_key,
-                                  const Slice& end_key) {
+std::optional<uint32_t> StressTest::GetRangeHash(
+    ThreadState* thread, const Snapshot* snapshot,
+    ColumnFamilyHandle* column_family, const Slice& start_key,
+    const Slice& end_key) {
   const std::string kCrcCalculatorSepearator = ";";
   uint32_t crc = 0;
   // This `ReadOptions` is for validation purposes. Ignore
@@ -2131,11 +2236,16 @@ uint32_t StressTest::GetRangeHash(ThreadState* thread, const Snapshot* snapshot,
     crc = crc32c::Extend(crc, kCrcCalculatorSepearator.data(), 1);
   }
   if (!it->status().ok()) {
-    fprintf(stderr, "Iterator non-OK when calculating range CRC: %s\n",
-            it->status().ToString().c_str());
-    thread->stats.AddErrors(1);
-    // Fail fast to preserve the DB state.
-    thread->shared->SetVerificationFailure();
+    if (it->status().IsMemoryLimit()) {
+      VerifyOneOpMemoryLimit(it->status(), thread);
+      return {};
+    } else {
+      fprintf(stderr, "Iterator non-OK when calculating range CRC: %s\n",
+              it->status().ToString().c_str());
+      thread->stats.AddErrors(1);
+      // Fail fast to preserve the DB state.
+      thread->shared->SetVerificationFailure();
+    }
   }
   return crc;
 }
@@ -2520,6 +2630,13 @@ void StressTest::Open(SharedState* shared) {
             continue;
           }
         }
+
+        if (s.IsMemoryLimit()) {
+          VerifyOneOpMemoryLimit(s, /*thread*/ nullptr);
+          printf("Retrying DB::Open after memory limit reached\n");
+          continue;
+        }
+
         break;
       }
     } else {